+++ tag/__init__.py
@@ -20,6 +20,7 @@ from tag.range import Range
 from tag.reader import GFF3Reader
 from tag.writer import GFF3Writer
 from tag.score import Score
+from tag import bae
 from tag import cli
 from tag import index
 from tag import locus